add _row_groups_to_parts utility
