+++ dask/dataframe/io/parquet/utils.py
@@ -1,5 +1,7 @@
 import re
 
+import pandas as pd
+
 from ....core import flatten
 
 
@@ -487,3 +489,144 @@ def _flatten_filters(filters):
         if filters
         else []
     )
+
+
+def _aggregate_stats(
+    file_path,
+    file_row_group_stats,
+    file_row_group_column_stats,
+    stat_col_indices,
+):
+    """Utility to aggregate the statistics for N row-groups
+    into a single dictionary.
+
+    Used by `Engine._construct_parts`
+    """
+    if len(file_row_group_stats) < 1:
+        # Empty statistics
+        return {}
+    elif len(file_row_group_column_stats) == 0:
+        assert len(file_row_group_stats) == 1
+        return file_row_group_stats[0]
+    else:
+        # Note: It would be better to avoid df_rgs and df_cols
+        #       construction altogether. It makes it fast to aggregate
+        #       the statistics for many row groups, but isn't
+        #       worthwhile for a small number of row groups.
+        if len(file_row_group_stats) > 1:
+            df_rgs = pd.DataFrame(file_row_group_stats)
+            s = {
+                "file_path_0": file_path,
+                "num-rows": df_rgs["num-rows"].sum(),
+                "total_byte_size": df_rgs["total_byte_size"].sum(),
+                "columns": [],
+            }
+        else:
+            s = {
+                "file_path_0": file_path,
+                "num-rows": file_row_group_stats[0]["num-rows"],
+                "total_byte_size": file_row_group_stats[0]["total_byte_size"],
+                "columns": [],
+            }
+
+        df_cols = None
+        if len(file_row_group_column_stats) > 1:
+            df_cols = pd.DataFrame(file_row_group_column_stats)
+        for ind, name in enumerate(stat_col_indices):
+            i = ind * 3
+            if df_cols is None:
+                s["columns"].append(
+                    {
+                        "name": name,
+                        "min": file_row_group_column_stats[0][i],
+                        "max": file_row_group_column_stats[0][i + 1],
+                        "null_count": file_row_group_column_stats[0][i + 2],
+                    }
+                )
+            else:
+                s["columns"].append(
+                    {
+                        "name": name,
+                        "min": df_cols.iloc[:, i].min(),
+                        "max": df_cols.iloc[:, i + 1].max(),
+                        "null_count": df_cols.iloc[:, i + 2].sum(),
+                    }
+                )
+        return s
+
+
+def _row_groups_to_parts(
+    fs,
+    gather_statistics,
+    split_row_groups,
+    file_row_groups,
+    file_row_group_stats,
+    file_row_group_column_stats,
+    stat_col_indices,
+    data_path,
+    partition_info,
+    make_part_split_func,
+    make_part_full_func,
+):
+
+    partition_keys = partition_info.get("partition_keys", None)
+    partition_obj = partition_info.get("partitions", None)
+
+    # Construct `parts` and `stats`
+    parts = []
+    stats = []
+    if split_row_groups:
+        # Create parts from each file,
+        # limiting the number of row_groups in each piece
+        split_row_groups = int(split_row_groups)
+        for filename, row_groups in file_row_groups.items():
+            row_group_count = len(row_groups)
+            for i in range(0, row_group_count, split_row_groups):
+                i_end = i + split_row_groups
+                rg_list = row_groups[i:i_end]
+                # Get full path (empty strings should be ignored)
+                full_path = fs.sep.join([p for p in [data_path, filename] if p != ""])
+
+                part = make_part_split_func(
+                    full_path,
+                    partition_keys,
+                    partition_obj,
+                    rg_list,
+                )
+                if part is None:
+                    continue
+
+                parts.append(part)
+                if gather_statistics:
+                    stat = _aggregate_stats(
+                        filename,
+                        file_row_group_stats[filename][i:i_end],
+                        file_row_group_column_stats[filename][i:i_end],
+                        stat_col_indices,
+                    )
+                    stats.append(stat)
+    else:
+        for filename, row_groups in file_row_groups.items():
+            # Get full path (empty strings should be ignored)
+            full_path = fs.sep.join([p for p in [data_path, filename] if p != ""])
+
+            part = make_part_full_func(
+                full_path,
+                partition_keys,
+                partition_obj,
+                row_groups,
+            )
+            if part is None:
+                continue
+
+            parts.append(part)
+            if gather_statistics:
+                stat = _aggregate_stats(
+                    filename,
+                    file_row_group_stats[filename],
+                    file_row_group_column_stats[filename],
+                    stat_col_indices,
+                )
+                stats.append(stat)
+
+    return parts, stats
